requirements.txt
discord.py[voice]
flask
Pillow
python-dotenv
requests